requirements.txt
# Production Dependencies for Smart Form Fill API
# Core FastAPI and web framework
fastapi==0.115.13
uvicorn==0.34.3
starlette==0.46.2
python-multipart==0.0.20

# Database and Redis
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
redis==6.2.0
hiredis==2.4.0

# OpenAI and LangChain (AI/ML)
openai==1.88.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langsmith==0.3.45
langgraph==0.4.8
tiktoken==0.9.0

# Vector operations and embeddings
numpy==2.3.0
faiss-cpu==1.11.0
scikit-learn==1.5.0
pinecone-client==3.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
docx2txt==0.9
chardet==5.2.0

# Authentication and security
PyJWT==2.10.1
passlib==1.7.4
bcrypt==4.3.0
python-ulid==1.1.0

# HTTP and requests
httpx==0.28.1
requests==2.32.4
aiohttp==3.12.13

# Data validation and serialization
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2

# Utilities
python-dotenv==1.1.0
loguru==0.7.3
tenacity==9.1.2

# Required dependencies (automatically pulled in)
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0
typing_extensions==4.14.0
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
click==8.2.1
packaging==24.2
annotated-types==0.7.0
greenlet==3.2.3
jiter==0.10.0
orjson==3.10.18
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
dataclasses-json==0.6.7
marshmallow==3.26.1
typing-inspect==0.9.0
typing-inspection==0.4.1
python-dateutil==2.9.0.post0
six==1.17.0
filelock==3.18.0
fsspec==2025.5.1
tqdm==4.67.1
networkx==3.5
Jinja2==3.1.6
MarkupSafe==3.0.2
distro==1.9.0
httpcore==1.0.9
httpx-sse==0.4.0
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.7.0
multidict==6.4.4
propcache==0.3.2
yarl==1.20.1
aiohappyeyeballs==2.6.1
requests-toolbelt==1.0.0
